+++ scripts/proxyArgs.ts
@@ -1,7 +1,0 @@
-const args = [
-  "0xB6C48B20c62F53C691335F313f1f2dE125993f0F", // implementation
-  "0x92371F3c1Bda1fB4F67a7984FDa54d60b0e6c6Ef", // initial owner
-  "0x8129fc1c" // init data
-];
-
-export default args;